+++ lcg.h
@@ -10,7 +10,7 @@ protected:
 	int a = 2147483629;
 	int b = 2147483587;
 public:
-	vector<long long> generate_random_numbers(int count, int seed);
+	vector<long long> generate_random_numbers(int count, long long seed);
 	long long mod_mult(long long a,long long b,long long mod);
 	long long power(long long x,long long y,long long m);
 	long long inv_mod(long long x,long long y,long long m);
